RouteInputAnalyzer now creates arrows dynamically
The arrows creation is based on "time to turn" which is just the
vehicles current distance/velocity. Based on this number the arrows
verticle location, horizontal size, and verticle size are dynamically
calculated. The base of the arrow is currently always located in the
middle of the screen. Additionally the arrow is always completely straight
which either points straight, left or right.